Fix wandering away from the end of the line when finding the points at the very beginning.
This was occurring because the check for the distance to the end of the sample line was not being activated until two points had been found. It was those two points which sometimes moved away from the end rather than towards it.

By changing the way the previous iteration's distance to the end of the line was calculated, then the first couple of points were included in the check, avoiding the possibility of moving away from the target position.