Add support for multiple segmented lines (i.e. profiles that bend).
To make this relatively simple, the parts of the function which found the nearest nodes have been moved into a new subfunction (__nodes_on_line__). The rest of the code is pretty much unchanged except it is now wrapped in a loop to do each section of the line.

Indices are stored in the order in which the points are given in positions (so make sure you get that right!).