Complete overhaul of the variable loading aspect of the class.
This is now much more streamlined (although I'm sure it could still be simplified) with a single function to load any variable of any shape.

self.load_timeseries() can now load a stride for the time dimensions, but I haven't implemented support for that in the main class. Manually calling that function can do that though.