Major change in how dimensions are handled.
We now handle all dimensions in the same way (i.e. time is not given as a start and end, but as a list of indices to extract, like all the others already were).

The old behaviour is a consequence on how the loading of data was structured. The new approach simply compares the variables' dimensions in the netCDF file with any which have been specified either at initialisation of the FileReader Object or when calling FileReader.load_data manually.

This is a breaking change, however!