Creating a Triangulation object and then using that to plot data on the elements breaks, whereas just passing the coordinates and the existing triangulation table to tripcolor works fine.
We'll do the latter so we can plot both element and node-based data with this method. However, this removes the masking bit. I'm not sure what that does anyway, so until someone moans, we'll just ignore it.

I'll leave it in despite this being in version control.

Also, squeeze out singleton dimensions from the data otherwise we frequently hit memory errors due to numpy broadcasting rules.