Complete change of approach for the namelist stuff.
Now, we create a dictionary whose keys are the NML_ sections and inside which is a list of NameListEntry objects. These have a name, value and type attribute which we use to write out to ASCII.

There is a new function to write out a dictionary of NML_ sections so the code can be reused for writing arbitrary namelists.