Pass file names instead of dataset objects to the various functions related to generating a FileReader object.
Since a Dataset object cannot be pickled, passing them around seems to leave them open in each of the called classes, making FileReader objects unpicklable. We rectify that here by simply opening the netCDF each time we need it. This is suboptimal (each open incurs a cost), but the penalty should be fairly minimal.