Invert the way we were doing things. So now, the canonical methods for arithmetic are the named functions (add, subtract, mulitply, divide and power) and the operators call those methods (so * calls multiply()).
The reason for doing it this way is that we can make the named methods have additional behaviour (namely working on specific variables only, for example) whilst not having to copy/paste code around between the named methods and operator approaches.

As it is there's a lot of boilerplate code in each named method which I'm not a big fan of, but that can be put into private methods if I get a moment.