Improve the performance of the subset node/element extraction.
Assume that we can load the entire grid into memory and just slice from that. The current set up works if you're severely memory constrained, but we're only loading the grid variables here, so if you're that short of memory, you're going to struggle to do any actual work.

The arrays are now loaded into memory and then dumped with the array indices.