Do a double check before declaring the triangulation invalid if we've been given nodes and elements.
The extra check identifies unique nodes in the elements asked for and compares that against the list of given nodes. If they match up, then the triangulation can be subset based on the given elements. If not, then we really have disparate sets of nodes and elements and there's no subsetting possible, in which case, just work as before.